Move triggers into a separate file
